pyproject.toml
[tool.poetry]
name = "pytorch-unet-segmentation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
torch = "^2.1.2"
scipy = "^1.11.4"
pillow = "^10.1.0"
matplotlib = "^3.8.2"
spikingjelly = "^0.0.0.0.14"
scikit-image = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
